Refactored Admin Schedule into Base Component for future reuse
